requirements.txt
absl-py==2.0.0
einops==0.7.0
numpy==1.24.4
pandas==2.0.3
spacy  # python -m spacy download en_core_web_sm
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
